Converted all the functions to use the new BadBlock Management (i think, may have missed a few...) Changed some variable types aswell... (nothing terribly important)

+++ src/x360Utils/NAND/NANDReader.cs
@@ -159,7 +159,7 @@
 
         public MobileEntry[] MobileArray { get; private set; }
 
-        public void SeekToLbaEx(ushort lba) {
+        public void SeekToLbaEx(uint lba) {
             if(_badBlocks.Contains(lba)) {
                 var block = 0;
                 while(true) {
